Create greetings.json
This file allows the exports and require functions to simplify what the exports from a string into a property; or from "Hola, hermano." to just greetings.es. This allows the Spanish greeting to be defined or even redefined elsewhere, while maintaining the greeting itself within app.js, outside this folder.